Commented out Console.WriteLine statements out of the MinAvgSliceTwiceWhatever project once again. Implemented a finite storage and an infinite storage revolving cyclic buffer based queue based on thoughts I had reading the reading material for Lesson 7: Stacks and Queues.

+++ Lessons/L5.PrefixSums.MinAvgTwoSlice/L5.PrefixSums.MinAvgTwoSlice/Solution.cs
@@ -33,7 +33,7 @@ namespace L5.PrefixSums.MinAvgTwoSlice
                 {
                     int n = j - i + 1;
                     long sumOfSlice = prefixSums[j + 1] - prefixSums[i];
-                    double average = sumOfSlice / n;
+                    double average = (double) sumOfSlice / n;
 
                     // Console.WriteLine($"slice ({i}, {j}), Average: {average}");
